fix: no split packages for compatibility with modules
